+++ src/com/marnia/world/tile/Tile.java
@@ -36,7 +36,7 @@ public class Tile {
 	public void getHitboxes(MarniaWorld world, int xt, int yt, List<AABB> hitboxes) {
 	}
 
-	public byte getSpriteData(MarniaWorld world, int xt, int yt) {
+	public int getSpriteData(MarniaWorld world, int xt, int yt) {
 		return 0;
 	}
 }
